Replaced snprintf with malloc + sprintf throughout. Fixed incorrect prototype for ceil_log2(). Removed prototype for (static) usage(). Changed #include "gmp.h" to #include <gmp.h> throughout.
git-svn-id: svn://scm.gforge.inria.fr/svnroot/ecm/trunk@812 404564d9-a503-0410-82bf-e18ce2cf3989

+++ auxlib.c
@@ -24,7 +24,7 @@
 #include <stdlib.h>
 #include <time.h>
 #include <stdarg.h>
-#include "gmp.h"
+#include <gmp.h>
 #include "ecm-impl.h"
 
 #define VERBOSE __ECM(verbose)
@@ -73,7 +73,7 @@ mpz_divby3_1op (mpz_t RS)
 
 /* returns ceil(log(n)/log(2)) */
 unsigned int
-ceil_log2 (unsigned int n)
+ceil_log2 (unsigned long n)
 {
   unsigned int k;
